refactor: 초기 prefetch 기능 제거
