+++ src/app/(afterLogin)/admin/content/themeCard/page.tsx
@@ -1,29 +1,10 @@
-import {
-  dehydrate,
-  HydrationBoundary,
-  QueryClient,
-} from '@tanstack/react-query';
-
 import Layout from '@/share/layout/Layout';
-import { usePrefetchCardList } from '@/share/query/card/useGetCardList';
-import { prefetchQuizList } from '@/share/query/quiz/useGetQuizList';
-import { prefetchVocaList } from '@/share/query/voca/useGetVocaList';
 
 import PageInfo from '../../_components/PageInfo';
 import CardList from './_components/CardList';
 import CreateCard from './_components/CreateCard';
 
 export default async function Card() {
-  const queryClient = new QueryClient();
-  await Promise.all([
-    usePrefetchCardList(queryClient),
-    prefetchVocaList(queryClient, {
-      data: { page: 1, page_size: 1000 },
-    }),
-    prefetchQuizList(queryClient, {
-      data: { page: 1, page_size: 1000 },
-    }),
-  ]);
   return (
     <>
       <Layout.Content.Full>
@@ -32,14 +13,12 @@ export default async function Card() {
           path={['콘텐츠', '테마별 카드모음집 관리']}
         />
       </Layout.Content.Full>
-      <HydrationBoundary state={dehydrate(queryClient)}>
-        <Layout.Content.Full>
-          <CardList />
-        </Layout.Content.Full>
-        <Layout.Content.Full>
-          <CreateCard />
-        </Layout.Content.Full>
-      </HydrationBoundary>
+      <Layout.Content.Full>
+        <CardList />
+      </Layout.Content.Full>
+      <Layout.Content.Full>
+        <CreateCard />
+      </Layout.Content.Full>
     </>
   );
 }
