+++ src/app/(afterLogin)/admin/quiz/quizManagement/[quizId]/page.tsx
@@ -1,14 +1,8 @@
-import {
-  dehydrate,
-  HydrationBoundary,
-  QueryClient,
-} from '@tanstack/react-query';
 import React from 'react';
 
 import CardFilter from '@/app/(afterLogin)/admin/quiz/_components/CardFilter';
 import QuizSearchList from '@/app/(afterLogin)/admin/quiz/quizManagement/_components/QuizSearchList';
 import Layout from '@/share/layout/Layout';
-import { prefetchQuizList } from '@/share/query/quiz/useGetQuizList';
 
 import PageInfo from '../../../_components/PageInfo';
 import UpdateQuiz from '../_components/UpdateQuiz';
@@ -18,25 +12,12 @@ export default async function QuizUpdate({
 }: {
   searchParams: { [key: string]: string | undefined };
 }) {
-  const queryClient = new QueryClient();
-
   const params = {
     start_date: searchParams['start_date'],
     end_date: searchParams['end_date'],
     is_use: searchParams['is_use'],
     search_keyword: searchParams['search_keyword'],
   };
-
-  await Promise.all([
-    prefetchQuizList(queryClient, {
-      data: {
-        page: 1,
-        page_size: 10,
-        sort_type: 'CREATED',
-        ...params,
-      },
-    }),
-  ]);
 
   return (
     <>
@@ -50,11 +31,9 @@ export default async function QuizUpdate({
         <CardFilter />
       </Layout.Content.Full>
 
-      <HydrationBoundary state={dehydrate(queryClient)}>
-        <Layout.Content.Left>
-          <QuizSearchList searchParams={params} />
-        </Layout.Content.Left>
-      </HydrationBoundary>
+      <Layout.Content.Left>
+        <QuizSearchList searchParams={params} />
+      </Layout.Content.Left>
       <Layout.Content.Right>
         <UpdateQuiz />
       </Layout.Content.Right>
