+++ src/app/(afterLogin)/admin/user/basicInfo/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
+import PageInfo from '@/app/(afterLogin)/admin/_components/PageInfo';
 import CardFilter from '@/app/(afterLogin)/admin/quiz/_components/CardFilter';
-import PageInfo from '@/app/(afterLogin)/admin/quiz/_components/PageInfo';
 import SegmentControlContainer from '@/app/(afterLogin)/admin/user/basicInfo/_components/SegmentControlContainer';
 import UserSearchList from '@/app/(afterLogin)/admin/user/basicInfo/_components/UserSearchList';
 import Layout from '@/share/layout/Layout';
